+++ R/nested_chart.R
@@ -34,6 +34,10 @@
 #'  (e.g. pi/2 = 90 degrees).
 #'@param subrotate_degrees integer; angle or vector of angles in degrees to
 #'  rotate the nested facet charts counter-clockwise by.
+#'@param zoom_x integer; vector with two values, the edges of the zoomed
+#'  section on the x-axis; defaults to NULL.
+#'@param zoom_y integer; vector with two values, the edges of the zoomed
+#'  section on the y-axis; defaults to NULL.
 #'@param file_width integer; file width in inches; defaults to 10.
 #'@param file_height integer; file height in inches; defaults to 10.
 #'@param dpi integer; resolution in dots per inch for "png" and "jpeg" files;
@@ -123,6 +127,13 @@
 #'  formats use \code{file_width}, \code{file_height}, and \code{dpi} to avoid
 #'  later rescaling and loss of quality.
 #'
+#'  Instead of using screenshots to crop the chart, it is highly recommendable
+#'  to use \code{zoom_x} and \code{zoom_y}. This allows for vector-based
+#'  graphics quality when showing sections of the chart. With this cropping
+#'  method, use \code{file_width} to set the overall size of the file output,
+#'  \code{file_height} will automatically adjust to retain the correct aspect
+#'  ratio, if both \code{zoom_x} and \code{zoom_y} are provided.
+#'
 #'  If \code{facet1} or \code{facet2} is \code{NA} for a given xarrow, the arrow
 #'  will end on the test's circle. Note: this correlation is usually not part of
 #'  the model.
@@ -188,6 +199,8 @@ nested_chart <- function(
   subrotate_degrees = 0,
   file_width = 10,
   file_height = 10,
+  zoom_x = NULL,
+  zoom_y = NULL,
   dpi = 500,
   color_global = "black",
   color_nested = "black",
@@ -249,6 +262,8 @@ nested_chart <- function(
     coord = coord,
     size = size,
     file_name = file_name,
+    zoom_x = zoom_x,
+    zoom_y = zoom_y,
     file_width = file_width,
     file_height = file_height,
     dpi = dpi,
@@ -889,7 +904,8 @@ coord_nested <- function (
                  cor_spacing     = cor_spacing,
                  arrows          = arrows)
   coord <- list(factor = factorcoords,
-                global = global)
+                global = global,
+                rs = rs)
   # if (prepare_item_charts == T) coord$items <- itemcoords
 
   return(coord)
@@ -906,6 +922,10 @@ coord_nested <- function (
 #'@param size integer; changes the size of most chart objects simultaneously.
 #'@param file_name character; name of the file to save. Supported formats are:
 #'  "pdf" (highest quality and smallest file size), "png", "jpeg"; defaults to "none".
+#'@param zoom_x integer; vector with two values, the edges of the zoomed
+#'  section on the x-axis; defaults to NULL.
+#'@param zoom_y integer; vector with two values, the edges of the zoomed
+#'  section on the y-axis; defaults to NULL.
 #'@param file_width integer; file width in inches; defaults to 10.
 #'@param file_height integer; file height in inches; defaults to 10.
 #'@param dpi integer; resolution in dots per inch for "png" and "jpeg" files;
@@ -956,6 +976,8 @@ plot_nested <- function (
   file_name = "none",
   file_width = 10,
   file_height = 10,
+  zoom_x = NULL,
+  zoom_y = NULL,
   dpi = 500,
   cor_labels_tests = TRUE,
   cor_labels_facets = TRUE,
@@ -1009,6 +1031,16 @@ plot_nested <- function (
     0.03 * size * sin(coord$g$axis_tick$phi) * coord$g$p_circs[1, "radius"]
 
   tick_label_label <- as.character(formatC(tick, format = "fg"))
+
+  # aspect ratio correction (to manage zoomed cases)
+  if(!is.null(zoom_x) & !is.null(zoom_y)) {
+    asp <- diff(zoom_y) / diff(zoom_x)
+    file_height <- asp * file_width
+  }
+
+  # scale zoom_x and zoom_y properly (messed up by relative_scaling)
+  if (!is.null(zoom_x)) zoom_x = zoom_x * coord$rs
+  if (!is.null(zoom_y)) zoom_y = zoom_y * coord$rs
 
 
   # chart ----------------------------------------------------------------------
@@ -1237,6 +1269,23 @@ plot_nested <- function (
   if (!is.null(title)) {
     myipv <- myipv +
       ggplot2::ggtitle(label = title)
+  }
+
+  # section
+  if (!is.null(c(zoom_x, zoom_y))) {
+    myipv <- myipv +
+      ggplot2::coord_cartesian(
+        xlim = zoom_x,
+        ylim = zoom_y,
+        expand = FALSE)
+    if(!is.null(zoom_x) & !is.null(zoom_y)) {
+      message(paste(
+        "file_height was set to ",
+        signif(asp, 4),
+        " times the file_width, to retain the aspect ratio.",
+        sep = ""))
+    }
+
   }